chore(cli): move constants to a separate file
moved all constants to a separate file for CLI scopre and imported from there.

ISSUES CLOSED: #772
