+++ bin/constants.js
@@ -1,0 +1,28 @@
+const NON_COMPILATION_ARGS = [
+	"init",
+	"migrate",
+	"add",
+	"remove",
+	"serve",
+	"generate-loader",
+	"generate-plugin",
+	"info"
+];
+
+const CONFIG_GROUP = "Config options:";
+const BASIC_GROUP = "Basic options:";
+const MODULE_GROUP = "Module options:";
+const OUTPUT_GROUP = "Output options:";
+const ADVANCED_GROUP = "Advanced options:";
+const RESOLVE_GROUP = "Resolving options:";
+const OPTIMIZE_GROUP = "Optimizing options:";
+const DISPLAY_GROUP = "Stats options:";
+const GROUPS = { CONFIG_GROUP, BASIC_GROUP, MODULE_GROUP, OUTPUT_GROUP, ADVANCED_GROUP, RESOLVE_GROUP, OPTIMIZE_GROUP, DISPLAY_GROUP }
+
+const webpackOptionsFlag = "WEBPACK_OPTIONS";
+
+module.exports = {
+  NON_COMPILATION_ARGS,
+  GROUPS,
+  webpackOptionsFlag,
+}
